refactor(automation): split github API by responsibility
Replace the monolithic github_api module with focused package modules for labels, issues, PRs, reviews, threads, checks, diffs, and issue state batching, with package export coverage.

Closes #1455

Implemented-By: Codex
