+++ hephaestus/automation/github_api/diff.py
@@ -1,0 +1,113 @@
+"""Pull-request diff position helpers."""
+
+from __future__ import annotations
+
+import re
+from typing import Any
+
+import hephaestus.automation.github_api as _api
+
+_HUNK_HEADER_RE = re.compile(r"^@@ -(\d+)(?:,\d+)? \+(\d+)(?:,\d+)? @@")
+
+
+def _valid_review_positions(diff_text: str) -> dict[str, set[tuple[int, str]]]:
+    """Map each changed file to the ``(line, side)`` positions GitHub will accept.
+
+    GitHub's review API rejects (HTTP 422) any inline comment whose ``line``/
+    ``side`` does not fall on a line present in the PR diff. A ``RIGHT`` comment
+    must target an added (``+``) or context (`` ``) line in the new file; a
+    ``LEFT`` comment must target a removed (``-``) or context line in the old
+    file. This parses the unified diff once into the set of accepted positions.
+
+    Args:
+        diff_text: Unified diff (``gh pr diff <n>`` output).
+
+    Returns:
+        ``{path: {(line_number, side), ...}}`` for every changed file.
+
+    """
+    positions: dict[str, set[tuple[int, str]]] = {}
+    current_path: str | None = None
+    old_line = 0
+    new_line = 0
+
+    for raw in diff_text.splitlines():
+        if raw.startswith("+++ "):
+            # ``+++ b/path`` (or ``+++ /dev/null``); strip the ``b/`` prefix.
+            target = raw[4:].strip()
+            if target == "/dev/null":
+                current_path = None
+            else:
+                current_path = target[2:] if target.startswith("b/") else target
+                positions.setdefault(current_path, set())
+            continue
+        if raw.startswith("--- "):
+            # Old-file header; new-file header (+++) sets the path.
+            continue
+
+        header = _HUNK_HEADER_RE.match(raw)
+        if header:
+            old_line = int(header.group(1))
+            new_line = int(header.group(2))
+            continue
+
+        if current_path is None or not raw:
+            continue
+
+        marker = raw[0]
+        if marker == "+":
+            positions[current_path].add((new_line, "RIGHT"))
+            new_line += 1
+        elif marker == "-":
+            positions[current_path].add((old_line, "LEFT"))
+            old_line += 1
+        elif marker == " ":
+            # Context line is valid on both sides.
+            positions[current_path].add((new_line, "RIGHT"))
+            positions[current_path].add((old_line, "LEFT"))
+            old_line += 1
+            new_line += 1
+        # Any other marker (e.g. ``\`` for "No newline") is ignored.
+
+    return positions
+
+
+def _filter_comments_to_diff(
+    comments: list[dict[str, Any]], diff_text: str
+) -> list[dict[str, Any]]:
+    """Drop inline comments whose ``(path, line, side)`` is not in the diff.
+
+    Prevents an out-of-hunk comment from making GitHub reject the *entire*
+    review with HTTP 422 (#1039). Dropped comments are logged at WARNING.
+
+    Fails open: if ``diff_text`` is empty (the diff could not be fetched), the
+    comments are returned unchanged — losing a comment because the diff was
+    unavailable would be worse than a possible 422.
+
+    Args:
+        comments: Inline comment dicts with ``path``/``line``/``side``/``body``.
+        diff_text: Unified diff to validate against.
+
+    Returns:
+        The subset of ``comments`` that target a line present in the diff.
+
+    """
+    if not diff_text.strip():
+        return comments
+
+    valid = _api._valid_review_positions(diff_text)
+    kept: list[dict[str, Any]] = []
+    for c in comments:
+        path = c.get("path", "")
+        line = c.get("line")
+        side = c.get("side", "RIGHT")
+        if path in valid and (line, side) in valid[path]:
+            kept.append(c)
+        else:
+            _api.logger.warning(
+                "Dropping out-of-hunk review comment on %s:%s (%s) — not in PR diff",
+                path,
+                line,
+                side,
+            )
+    return kept
